+++ config/settings.py
@@ -38,7 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'chat',
+    'retrieval_qa_with_source',
 ]
 
 MIDDLEWARE = [
