+++ retrieval_qa_with_source/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ChatConfig(AppConfig):
+class RetrievalQaWithSourceConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'chat'
+    name = 'retrieval_qa_with_source'
