+++ retrieval_qa_with_source/models.py
@@ -2,7 +2,7 @@ from django.contrib.auth.models import User
 from django.db import models
 
 
-class ChatGpt(models.Model):
+class ChatLogs(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     thread = models.CharField(max_length=255)
     role = models.CharField(max_length=255)
