+++ retrieval_qa_with_source/templates/retrieval_qa_with_source/home.html
@@ -1,27 +1,27 @@
-{% extends 'chat/base.html' %}
+{% extends 'retrieval_qa_with_source/base.html' %}
 {% load static %}
 
 {% block content %}
     <div class="container">
         <div class="jumbotron">
-            <h1 class="display-4">ChatGPT</h1>
+            <h1 class="display-4">ChatGPT with source</h1>
             <p class="lead">The system that forms the basis of work</p>
             <hr class="my-4">
-            <p>You can issue ChatGPT.</p>
+            <p>You can talk with ChatGPT.</p>
         </div>
 
-        {% for chatlog in chatlogs %}
+        {% for chat_log in chat_logs %}
             <div class="card mb-3">
                 <div class="card-body">
-                    <h5 class="card-title">{{ chatlog.role }}</h5>
-                    <h6 class="card-subtitle mb-2 text-muted">{{ chatlog.thread }}</h6>
-                    <p class="card-text">{{ chatlog.message }}</p>
+                    <h5 class="card-title">{{ chat_log.role }}</h5>
+                    <h6 class="card-subtitle mb-2 text-muted">{{ chat_log.thread }}</h6>
+                    <p class="card-text">{{ chat_log.message }}</p>
                     <a href="#" class="card-link">Card link</a>
                     <a href="#" class="card-link">Another link</a>
                 </div>
             </div>
         {% endfor %}
-        <form action="{% url 'cht:home' %}" method="POST">
+        <form action="{% url 'qa_with_src:home' %}" method="POST">
             {{ form }}
             {% csrf_token %}
             <input type="submit" value="送信">
