+++ retrieval_qa_with_source/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 from . import views
 
-app_name = 'cht'
+app_name = 'qa_with_src'
 urlpatterns = [
     path('', views.HomeView.as_view(), name='home'),
 ]
